// backend/src/common/enums/roles.enum.ts
export enum UserRoleEnum {
  SUPER_ADMIN = 'SUPER_ADMIN',
  DISPATCHER = 'DISPATCHER',
  DRIVER = 'DRIVER',
  PASSENGER = 'PASSENGER',
}

export enum UserStatusEnum {
  PENDING_APPROVAL = 'PENDING_APPROVAL',
  ACTIVE = 'ACTIVE',
  SUSPENDED = 'SUSPENDED',
  INACTIVE = 'INACTIVE',
}

export enum VehicleCategoryEnum {
  EXECUTIVE_SEDAN = 'EXECUTIVE_SEDAN',
  VIP_SUV = 'VIP_SUV',
  PREMIUM_VAN = 'PREMIUM_VAN',
  LUXURY_ARMORED = 'LUXURY_ARMORED',
}

export enum VehicleStatusEnum {
  AVAILABLE = 'AVAILABLE',
  IN_SERVICE = 'IN_SERVICE',
  MAINTENANCE = 'MAINTENANCE',
  DECOMMISSIONED = 'DECOMMISSIONED',
}

export enum DocumentTypeEnum {
  DRIVER_LICENSE = 'DRIVER_LICENSE',
  CRIMINAL_RECORD = 'CRIMINAL_RECORD',
  IDENTITY_CARD = 'IDENTITY_CARD',
  MEDICAL_CERTIFICATE = 'MEDICAL_CERTIFICATE',
  DRIVING_CERTIFICATE = 'DRIVING_CERTIFICATE',
  VEHICLE_TITLE = 'VEHICLE_TITLE',
  SOAT_INSURANCE = 'SOAT_INSURANCE',
  TECHNICAL_INSPECTION = 'TECHNICAL_INSPECTION',
}

export enum DocumentStatusEnum {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  EXPIRED = 'EXPIRED',
}

export enum ShiftStatusEnum {
  ACTIVE = 'ACTIVE',
  COMPLETED = 'COMPLETED',
  FORCED_CLOSED = 'FORCED_CLOSED',
}

export enum RideStatusEnum {
  SOLICITADO = 'SOLICITADO',
  ASIGNADO = 'ASIGNADO',
  EN_CAMINO = 'EN_CAMINO',
  ABORDAJE = 'ABORDAJE',
  EN_CURSO = 'EN_CURSO',
  FINALIZADO = 'FINALIZADO',
  CANCELADO = 'CANCELADO',
}